refactor(access): move in-memory checker
